로그인 기능 완료, 권한 부혀 및 기타작업 완료

+++ src/main/java/com/practiceprj1/controller/BoardController.java
@@ -3,6 +3,8 @@ package com.practiceprj1.controller;
 import com.practiceprj1.domain.Board;
 import com.practiceprj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +19,7 @@ import java.util.List;
 public class BoardController {
     private final BoardService service;
 
+
     @GetMapping("add")
     public String add() {
 
@@ -24,10 +27,11 @@ public class BoardController {
     }
 
     @PostMapping("add")
-    public String add(Board board) {
-        service.add(board);
+    public String add(Board board, Authentication authentication, RedirectAttributes rttr) {
+        service.add(board, authentication);
 
-        return "redirect:/board/add";
+        rttr.addAttribute("id", board.getId());
+        return "redirect:/board";
     }
 
     @GetMapping("")
@@ -57,16 +61,20 @@ public class BoardController {
     }
 
     @PostMapping("modify")
-    public String modify(Board board, RedirectAttributes rttr) {
-        service.update(board);
-        rttr.addAttribute("id", board.getId());
+    public String modify(Board board, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.update(board);
+            rttr.addAttribute("id", board.getId());
+        }
 
         return "redirect:/board";
     }
 
     @PostMapping("delete")
-    public String delete(Integer id) {
-        service.delete(id);
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.delete(id);
+        }
 
         return "redirect:/";
     }
